Did more work on routing, the app bar is now showing up everywehere regardless of the page
